% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation.pos.utils}
\alias{annotation.pos.utils}
\title{A position annotation utils that can be used to write a yourself annotation function}
\usage{
annotation.pos.utils(dat = data.table(), name = "", builder = "hg19",
  database.dir = Sys.getenv("annovarR_DB_DIR", NULL), db.col.order = 1:5,
  index.col = c("chr", "start"), matched.cols = c("chr", "start", "end",
  "ref", "alt"), return.col.index = 6, return.col.names = "",
  format.dat.fun = format.cols, setdb.fun = set.db,
  set.table.fun = set.table, format.db.tb.fun = format.db.tb,
  db.type = "sqlite", verbose = FALSE)
}
\arguments{
\item{dat}{A data.table including all of your data, eg. data.table(chr=c(1,2,3), start=c(1111,1112,1113))}

\item{name}{Annotation name, eg. avsnp138, avsnp147, 1000g2015aug_all}

\item{builder}{Genome version, hg19, hg38, mm10 and others}

\item{database.dir}{Dir of the databases}

\item{db.col.order}{Using the index, you can rename the database table, and can be matched using matched.cols.}

\item{index.col}{Using the selected cols to match data with sqlite database. eg. c('chr', 'start'), 'rs'}

\item{matched.cols}{Using the selected cols to match data with selected partial data by index.col limited.}

\item{return.col.index}{Setting the colnums need be returned}

\item{return.col.names}{Setting the returned colnum names}

\item{format.dat.fun}{A function to process input data. eg. as.numeric(dat$start); as.character(dat$chr)}

\item{setdb.fun}{A function to process the name, builder, database.dir and get the database path}

\item{set.table.fun}{A function to process the name, builder and get the final table name}

\item{format.db.tb.fun}{A function to process the selected database table that can be used to matched with your data}

\item{db.type}{Setting the database type (sqlite or txt)}

\item{verbose}{Logical indicating wheather print the extra log infomation}
}
\description{
A position annotation utils that can be used to write a yourself annotation function
}
\examples{
library(data.table)
chr <- c('chr1', 'chr2', 'chr1')
start <- c('10020', '10020', '10020')
end <- c('10020', '10020', '10020')
ref <- c('A', 'A', 'A')
alt <- c('-', '-', '-')
database <- system.file('extdata', 'demo/hg19_avsnp147.sqlite', package = 'annovarR')
database.dir <- dirname(database)
dat <- data.table(chr = chr, start = start, end = end, ref = ref, alt = alt)
x <- annotation.pos.utils(dat, 'avsnp147', database.dir = database.dir, 
return.col.names = 'avSNP147')
}
