% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\title{Annotation function}
\usage{
annotation()
}
\arguments{
\item{dat.list}{A list including all of your data, eg. list(chr=c(1,2,3), start=c(1111,1112,1113))}

\item{name}{Annotation name, eg. avsnp138, avsnp147, 1000g2015aug_all}

\item{builder}{Genome version, hg19, hg38, mm10 and others}

\item{verbose}{Logical indicating wheather print the extra log infomation}
}
\description{
Annotation function
}
\examples{
chr <- c('chr1', 'chr2', 'chr1')
start <- c('10020', '10020', '10020')
end <- c('10020', '10020', '10020')
ref <- c('A', 'A', 'A')
alt <- c('-', '-', '-')
database <- system.file('extdata', 'hg19_avsnp147.sqlite', package = 'annovarR')
database.dir <- dirname(database)
dat.list <- list(chr = chr, start = start, end = end, ref = ref, alt = alt)
x <- annotation(dat.list, 'avsnp147', database.dir = database.dir, return.col.names = 'avSNP147')
}
