% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_tools.R
\name{annovar}
\alias{annovar}
\title{R utils function to run ANNOVAR tool.}
\usage{
annovar(perl = Sys.which("perl"), cmd.pool = list(script1.downdb =
  "{{perl}} {{script}}{{extra.params}} -downdb {{buildver}} {{webfrom}} {{down.dbname}} {{database.dir}}",
  script1.gene.based =
  "{{perl}} {{script}}{{extra.params}} {{buildver}} {{input.file}} {{database.dir}}",
  script1.region.based =
  "{{perl}} {{script}} -regionanno{{extra.params}} {{buildver}} {{dbtype}} {{input.file}} {{database.dir}}",
  script1.filter.based =
  "{{perl}} {{script}} -filter{{extra.params}} {{buildver}} {{dbtype}} {{input.file}} {{database.dir}}",
  script2 =
  "{{perl}} {{script}} {{input.file}} {{database.dir}} {{buildver}} -remove -protocol {{dbtype}} -operation {{operation}} -nastring . -otherinfo -vcfinput"),
  cmd.used = "script1.downdb", down.dbname = "", input.file = "",
  annovar.dir = "", buildver = "hg19",
  database.dir = "{{annovar.dir}}/humandb", webfrom = "annovar",
  dbtype = "", out = NULL, operation.type = list(gene.based = c("refGene",
  "knownGene", "ensGene", "ccdsGene"), region.based = c("cytoBand",
  "genomicSuperDups")), cmd.profix.flag = list(buildver = "-buildver", dbtype
  = "-dbtype", webfrom = "-webfrom", out = "-out"), extra.params = "",
  debug = FALSE)
}
\arguments{
\item{perl}{Executable file of perl}

\item{cmd.pool}{Un-parsed commands of ANNOVAR}

\item{cmd.used}{Name in cmd.pool that used to parse final run command}

\item{down.dbname}{Need to download database name, e.g. avsnp147,cosmic70,1000g2015aug}

\item{input.file}{Input file name, e.g. example.avinput, example.vcf}

\item{annovar.dir}{ANNOVAR source code directory}

\item{buildver}{Genome version e.g. hg19, mm10}

\item{database.dir}{Database directory, e.g. /opt/annovar/humandb}

\item{webfrom}{Database resource warehouse, e.g. ucsc, annovar}

\item{dbtype}{ANNOVAR annotation name}

\item{out}{ANNOVAR -out parameter value}

\item{operation.type}{Operation types used in table_annovar.pl}

\item{cmd.profix.flag}{Profix used in ANNOVAR command}

\item{extra.params}{Extra paramters in ANNOVAR command}

\item{debug}{If set TRUE, only print the command}
}
\description{
R utils function to run ANNOVAR tool.
}
\examples{
down.dbname <- 'refGene'
annovar('perl', down.dbname = 'avsnp147', annovar.dir = '/opt/annovar', debug = TRUE)
}
