% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovarR.R
\docType{package}
\name{annovarR}
\alias{annovarR}
\alias{annovarR-package}
\title{annovarR is an integrated database and tool to annotate genetic variants 
from genomic and transcriptome sequencing data.
This work mainly based on ANNOVAR and other public annotation databases.}
\description{
annovarR is an integrated database and tool to annotate genetic variants 
from genomic and transcriptome sequencing data.
This work mainly based on ANNOVAR and other public annotation databases.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/annovarR} 

Report bugs at \url{https://github.com/JhuangLab/annovarR/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
