% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert2annovar}
\alias{convert2annovar}
\title{According the ANNOVAR convert2annovar.pl script, rewrite using R}
\usage{
convert2annovar(input.format = "pileup", out.file = NULL, snp.qual = 20,
  snp.pvalue = 1, coverage = 0, max.coverage = NULL,
  include.info = NULL, chr = "", chrmt = "M", alt.cov = NULL,
  allelic.frac = NULL, fraction = NULL, species = "human",
  filter.output = NULL, all.sample = NULL, with.zyg = NULL,
  genotype.qual = NULL, variant.qual = NULL, comment = NULL,
  dbsnp.file = NULL, with.freq = NULL, seq.dir = NULL, ins.size = NULL,
  del.size = NULL, sub.size = NULL, context = NULL, verbose = FALSE)
}
\arguments{
\item{input.format}{format (default: pileup)}

\item{out.file}{Output file name (default: STDOUT)}

\item{snp.qual}{Quality score threshold in pileup file (default: 20)}

\item{snp.pvalue}{P-value threshold in GFF3-SOLiD file (default: 1)}

\item{coverage}{Read coverage threshold in pileup file (default: 0)}

\item{max.coverage}{Maximum coverage threshold (default: none)}

\item{include.info}{Include supporting information in output}

\item{chr}{Specify the chromosome (for CASAVA format)}

\item{chrmt}{chr identifier for mitochondria (default: M)}

\item{alt.cov}{Alternative allele coverage threshold (for pileup format)}

\item{allelic.frac}{Print out allelic fraction rather than het/hom status (for pileup format)}

\item{fraction}{Minimum allelic fraction to claim a mutation (for pileup format)}

\item{species}{If human, Convert chr23/24/25 to X/Y/M (for gff3-solid format)}

\item{filter.output}{Variants with this filter (case insensitive, for vcf4 format)}

\item{all.sample}{Process all samples in file with separate output files (for vcf4 format)}

\item{with.zyg}{print Zygosity/coverage/quality when -includeinfo is used (for vcf4 format)}

\item{genotype.qual}{Genotype quality score threshold (for vcf4 format)}

\item{variant.qual}{Variant quality score threshold (for vcf4 format)}

\item{comment}{Keep comment line in output (for vcf4 format)}

\item{dbsnp.file}{dbSNP file in UCSC format (for rsid format)}

\item{with.freq}{For --allsample, print frequency information instead (for vcf4 format)}

\item{seq.dir}{Directory with FASTA sequences (for region format)}

\item{ins.size}{Insertion size (for region format)}

\item{del.size}{Deletion size (for region format)}

\item{sub.size}{Substitution size (default: 1, for region format)}

\item{context}{Print context nucleotide for indels (for casava format)}

\item{verbose}{Ligical indicating wheather show the log message}
}
\description{
According the ANNOVAR convert2annovar.pl script, rewrite using R
}
\examples{
example.vcf <- system.file('extdata', 'demo/example.vcf', package = 'annovarR')
}
