% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.database}
\alias{download.database}
\title{Download annovarR databases}
\usage{
download.database(catgry.name = c(), version = c(), buildver = "hg19",
  database.dir = c(), database.cfg = system.file("extdata",
  "config/db/db_annovar.toml", package = "BioInstaller"),
  show.all.versions = FALSE, show.all.names = FALSE,
  show.all.buildvers = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{catgry.name}{Name of download catgry eg. cosmic, avsnp, 1000g}

\item{version}{Version of download database}

\item{buildver}{Genome version, e.g hg19, hg38, mm10}

\item{database.dir}{Dir of the databases}

\item{database.cfg}{Configuration file for download, most of annovarR database 
resources can be found in system.file('extdata', 
'config/db/db_annovar.toml', package = "BioInstaller") and 
system.file('extdata', 'config/db/db_main.toml', package = "BioInstaller")}

\item{show.all.versions}{Logical indicating wheather show all avaliable versions}

\item{show.all.names}{Logical indicating wheather show all avaliable name}

\item{show.all.buildvers}{Logical indicating wheather show all avaliable buildver}

\item{verbose}{Logical indicating wheather print the extra log infomation}

\item{...}{Other parameters pass to \code{\link[BioInstaller]{install.bioinfo}}}
}
\description{
Download annovarR databases
}
\examples{
download.database('db_annovar_1000g', database.dir = sprintf('\%s/databases/', tempdir()), 
show.all.versions = TRUE)

download.database(show.all.names = TRUE)
download.database(database.cfg = system.file('extdata', 'config/db/db_main.toml', package = "BioInstaller"), 
show.all.names = TRUE)
}
