% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{drop.mysql.index}
\alias{drop.mysql.index}
\title{Drop search index of annovarR database table in mysql}
\usage{
drop.mysql.index(mysql.connect.params = list(host = "", dbname = "",
  table.name = "", user = "user", password = "password"), index = "",
  verbose = FALSE, ...)
}
\arguments{
\item{mysql.connect.params}{Mysql parameters, [host, dbname, table.name, user, password etc.]}

\item{index}{Index name in mysql}

\item{verbose}{Ligical indicating wheather show the log message}

\item{...}{Other parameters be used in dbSendQuery}
}
\description{
Drop search index of annovarR database table in mysql
}
\examples{
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
\dontrun{
drop.mysql.index(list(host = 'host', dbname = 'db', user = 'user', password = 'password'), 
index = 'index_name')
}
}
