% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{drop.sqlite.index}
\alias{drop.sqlite.index}
\title{Drop search index of annovarR database table in sqlite}
\usage{
drop.sqlite.index(db, index, verbose = FALSE)
}
\arguments{
\item{db}{Output path of sqlite database}

\item{index}{Index name in sqlite}

\item{verbose}{Ligical indicating wheather show the log message}
}
\description{
Drop search index of annovarR database table in sqlite
}
\examples{
test.sqlite <- sprintf('\%s/snp.test.sqlite', tempdir())
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
x <- sqlite.build(filename = test.dat, db = test.sqlite, table.name = 'snp_test')
x <- sqlite.index(db = test.sqlite, table.name = 'snp_test', index = 'index4', cols = c('V1', 'V2'))
x <- drop.sqlite.index(db = test.sqlite, index = 'index4')
unlink(test.sqlite)
}
