% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.annotation.dbtype}
\alias{get.annotation.dbtype}
\title{Get annovarR default databases type [sqlite, txt]}
\usage{
get.annotation.dbtype(anno.name = "", database.cfg = system.file("extdata",
  "config/databases.toml", package = "annovarR"))
}
\arguments{
\item{anno.name}{Annotation name, eg. avsnp138, avsnp147, 1000g2015aug_all}

\item{database.cfg}{Configuration file of annovarR databases infomation}
}
\description{
Get annovarR default databases type [sqlite, txt]
}
\examples{
get.annotation.dbtype('avsnp147') 
}
