% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.annotation.names}
\alias{get.annotation.names}
\title{Get all anor supported databases}
\usage{
get.annotation.names(
  database.cfg = system.file("extdata", "config/databases.toml", package = "anor")
)
}
\arguments{
\item{database.cfg}{Configuration file of anor databases infomation}
}
\description{
Get all anor supported databases
}
\examples{
cfg <- system.file('extdata', 'config/config.toml', package = 'anor')
get.annotation.names(cfg) 
}
