% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.annotation.needcols}
\alias{get.annotation.needcols}
\title{Get annovarR annotation needed colnames according the anno.name}
\usage{
get.annotation.needcols(anno.name = "",
  database.cfg = system.file("extdata", "config/databases.toml", package =
  "annovarR"))
}
\arguments{
\item{anno.name}{Annotation name, eg. avsnp138, avsnp147, 1000g2015aug_all}

\item{database.cfg}{Configuration file of annovarR databases infomation}
}
\description{
Get annovarR annotation needed colnames according the anno.name
}
\examples{
get.annotation.dbtype('avsnp147') 
}
