% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get.download.name}
\alias{get.download.name}
\title{Use annotation name to get download.name that can be used 
to download the database use \code{download.database}}
\usage{
get.download.name(anno.name = "", database.cfg = system.file("extdata",
  "config/databases.toml", package = "annovarR"))
}
\arguments{
\item{anno.name}{Annotation name, eg. avsnp138, avsnp147, 1000g2015aug_all}

\item{database.cfg}{Configuration file of annovarR databases infomation}
}
\description{
Use annotation name to get download.name that can be used 
to download the database use \code{download.database}
}
\examples{
get.download.name('avsnp147')
}
