% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{mysql.head}
\alias{mysql.head}
\title{Show top n line of table of database in mysql database}
\usage{
mysql.head(mysql.connect.params = list(host = "", dbname = "", table.name =
  "", user = "", password = ""), n = 10, extra.sql = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{mysql.connect.params}{Mysql parameters, [host, dbname, table.name, user, password etc.]}

\item{n}{n lines will be selected}

\item{extra.sql}{Extra sql statement}

\item{verbose}{Ligical indicating wheather show the log message}

\item{...}{Other parameters pass to dbGetQuery}
}
\description{
Show top n line of table of database in mysql database
}
\examples{
host <- '11.11.11.1'
}
