% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{mysql.index}
\alias{mysql.index}
\title{Generate search index of annovarR database table in mysql}
\usage{
mysql.index(mysql.connect.params = list(host = "", dbname = "", table.name =
  "", user = "user", password = "password"), index = "", cols = c(),
  verbose = FALSE, ...)
}
\arguments{
\item{mysql.connect.params}{Mysql parameters, [host, dbname, table.name, user, password etc.]}

\item{index}{Index name in mysql}

\item{cols}{Colnames needed with length to be index (e.g. c('V1(6)', 'V2'))}

\item{verbose}{Ligical indicating wheather show the log message}

\item{...}{Other parameters be used in dbSendQuery}
}
\description{
Generate search index of annovarR database table in mysql
}
\examples{
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
\dontrun{
mysql.index(list(host = 'host', dbname = 'db', table.name = 'table', 
user = 'user', password = 'password'), index = 'index_name', cols = c('V1', 'V2'))
}
}
