% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{mysql.tb.colnames}
\alias{mysql.tb.colnames}
\title{Get colnames of table of database in mysql}
\usage{
mysql.tb.colnames(mysql.connect.params = list(host = "", dbname = "", user =
  "", password = "", table.name = ""), ...)
}
\arguments{
\item{mysql.connect.params}{Mysql parameters, [host, dbname, table.name, user, password etc.]}

\item{...}{Other parameters pass to dbConnect}
}
\description{
Get colnames of table of database in mysql
}
\examples{
\dontrun{
mysql.db.colnames(list(host = 'host', dbname = 'db', user = 'user', 
password = 'password', table.name = 'table'))
}
}
