% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{mysql.tb.indexes}
\alias{mysql.tb.indexes}
\title{Get mysql table indexes}
\usage{
mysql.tb.indexes(mysql.connect.params = list(dbname = "", table.name = ""))
}
\arguments{
\item{mysql.connect.params}{Connect to mysql database params [dbname, table.name, host, user, password]}
}
\description{
Get mysql table indexes
}
\examples{
NULL
}
