% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parAnnotation}
\alias{parAnnotation}
\title{Annotation function (mulitple name)}
\usage{
parAnnotation(row.cl, ...)
}
\arguments{
\item{row.cl}{Cluster object be generated by makeCluster}

\item{...}{Other parametes see \code{\link{annotation.merge}}}
}
\description{
Annotation function (mulitple name)
}
\examples{
library(data.table)
chr <- c('chr1', 'chr2', 'chr1')
start <- c('10020', '10020', '10020')
end <- c('10020', '10020', '10020')
ref <- c('A', 'A', 'A')
alt <- c('-', '-', '-')
database <- system.file('extdata', 'demo/hg19_avsnp147.txt', package = 'annovarR')
database.dir <- dirname(database)
dat <- data.table(chr = chr, start = start, end = end, ref = ref, alt = alt)
library(parallel)
row.cl <- makeCluster(2)
x <- parAnnotation(dat = dat, anno.names = c('avsnp147', '1000g2015aug_all'), 
database.dir = database.dir, db.type = 'txt', row.cl = row.cl)
stopCluster(row.cl)
}
