% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{show.cfg.databses}
\alias{show.cfg.databses}
\title{Show all annovarR supported databases}
\usage{
show.cfg.databses(cfg = NULL)
}
\arguments{
\item{cfg}{annovarR configuration file}
}
\description{
Show all annovarR supported databases
}
\examples{
cfg <- system.file('extdata', 'config/config.toml', package = 'annovarR')
show.cfg.databses() 
}
