% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{sqlite.build}
\alias{sqlite.build}
\title{Build annovarR database in sqlite}
\usage{
sqlite.build(filename, db, table.name, new.colnames = NULL,
  overwrite = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{filename}{Path of raw data}

\item{db}{Output path of sqlite database}

\item{table.name}{Table name in sqlite}

\item{new.colnames}{New colnames of table, default is to retain the original}

\item{overwrite}{Ligical indicating wheather overwrite sqlite database}

\item{verbose}{Ligical indicating wheather show the log message}

\item{...}{Other parameters be used in \code{\link[data.table]{fread}}}
}
\description{
Build annovarR database in sqlite
}
\examples{
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
test.sqlite <- sprintf('\%s/snp.test.sqlite', tempdir())
x <- sqlite.build(filename = test.dat, db = test.sqlite, 'snp_test')
unlink(test.sqlite)
}
