% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sqlite.head}
\alias{sqlite.head}
\title{Show top n line of table of database in sqlite}
\usage{
sqlite.head(db, table.name, n = 10)
}
\arguments{
\item{db}{Output path of sqlite database}

\item{table.name}{Table name in sqlite}

\item{n}{Top n lines will be selected}
}
\description{
Show top n line of table of database in sqlite
}
\examples{
test.sqlite <- sprintf('\%s/snp.test.sqlite', tempdir())
test.dat <- system.file('extdata', 'sqlite.dat.txt', package = 'annovarR')
x <- sqlite.build(filename = test.dat, db = test.sqlite, table.name = 'snp_test')
sqlite.head(test.sqlite, "snp_test")
}
