% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sqlite.tb.colnames}
\alias{sqlite.tb.colnames}
\title{Get colnames of table of database in sqlite}
\usage{
sqlite.tb.colnames(db.path, table.name)
}
\arguments{
\item{db.path}{Path of sqlite database}

\item{table.name}{Table name in sqlite}
}
\description{
Get colnames of table of database in sqlite
}
\examples{
test.sqlite <- sprintf('\%s/snp.test.sqlite', tempdir())
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
x <- sqlite.build(filename = test.dat, db = test.sqlite, table.name = 'snp_test')
sqlite.tb.colnames(test.sqlite, 'snp_test')
}
