% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\title{Function to run annovarR shiny APP service}
\usage{
web(appDir = system.file("extdata", "tools/shiny/R", package = "annovarR"),
  ...)
}
\arguments{
\item{appDir}{The application to run.
Default is system.file('extdata', 'tools/shiny/R', package = 'annovarR')}

\item{...}{Other parameters pass to \code{\link[shiny]{runApp}}}
}
\description{
Function to run annovarR shiny APP service
}
\examples{
\dontrun{
  web()
}
}
