% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovarR.R
\docType{package}
\name{annovarR}
\alias{annovarR}
\alias{annovarR-package}
\title{annovarR is an efficient software tool to utilize update-to-date information to 
functionally annotate genetic variants detected from diverse genomes based on ANNOVAR.}
\description{
annovarR is an efficient software tool to utilize update-to-date information to 
functionally annotate genetic variants detected from diverse genomes based on ANNOVAR.
}
\seealso{
Useful links:

\url{https://github.com/Miachol/annovarR} 

Report bugs at \url{https://github.com/Miachol/annovarR/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
