% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovarR.R
\docType{package}
\name{annovarR}
\alias{annovarR}
\alias{annovarR-package}
\title{annovarR is an novel integrated open source tool to annotate genetic variants data 
based on ANNOVAR and other public annotation databases, 
such as [varcards](http://varcards.biols.ac.cn/), 
[REDIportal](http://srv00.recas.ba.infn.it/atlas/), .etc.}
\description{
annovarR is an novel integrated open source tool to annotate genetic variants data 
based on ANNOVAR and other public annotation databases, 
such as [varcards](http://varcards.biols.ac.cn/), 
[REDIportal](http://srv00.recas.ba.infn.it/atlas/), .etc.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/annovarR} 

Report bugs at \url{https://github.com/JhuangLab/annovarR/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
