% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovarR.R
\docType{package}
\name{annovarR}
\alias{annovarR}
\alias{annovarR-package}
\title{annovarR can be used to annotate genetic variants from genomic and transcriptome data. 
Some of published and in-house database have been introduced in annovarR.}
\description{
annovarR can be used to annotate genetic variants from genomic and transcriptome data. 
Some of published and in-house database have been introduced in annovarR.
}
\seealso{
Useful links:

\url{https://github.com/JhuangLab/annovarR} 

Report bugs at \url{https://github.com/JhuangLab/annovarR/issues}
}
\author{
Li Jianfeng \url{lee_jianfeng@sjtu.edu.cn}
}
