% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.database}
\alias{download.database}
\title{Download annovarR databases}
\usage{
download.database(name = c(), version = c(), database.dir = c(),
  database.cfg = system.file("extdata", "config/download.toml", package =
  "annovarR"), verbose = FALSE, ...)
}
\arguments{
\item{name}{Name of download eg. cosmic, avsnp, 1000g}

\item{version}{Version of download database}

\item{database.dir}{Dir of the databases}

\item{database.cfg}{Configuration file of annovarR databases infomation}

\item{verbose}{Logical indicating wheather print the extra log infomation}

\item{...}{Other parameters pass to \code{\link[BioInstaller]{install.bioinfo}}}
}
\description{
Download annovarR databases
}
\examples{
download.database('1000g', database.dir = sprintf('\%s/databases/', tempdir()), show.all.versions = TRUE)
}
