% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get.annotation.names}
\alias{get.annotation.names}
\title{Get all annovarR supported databases}
\usage{
get.annotation.names(database.cfg = system.file("extdata",
  "config/databases.toml", package = "annovarR"))
}
\arguments{
\item{database.cfg}{Configuration file of annovarR databases infomation}
}
\description{
Get all annovarR supported databases
}
\examples{
cfg <- system.file('extdata', 'config/config.toml', package = 'annovarR')
get.annotation.names(cfg) 
}
