% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sqlite.head}
\alias{sqlite.head}
\title{Show top n line of table of database in sqlite database}
\usage{
sqlite.head(sqlite.connect.params = list(dbname = "", table.name = ""),
  n = 10, extra.sql = NULL, verbose = FALSE, ...)
}
\arguments{
\item{sqlite.connect.params}{Connect to sqlite database params [dbname, table.name]}

\item{n}{n lines will be selected}

\item{extra.sql}{Extra sql statement}

\item{verbose}{Ligical indicating wheather show the log message}

\item{...}{Other parameters pass to dbGetQuery}
}
\description{
Show top n line of table of database in sqlite database
}
\examples{
test.sqlite <- sprintf('\%s/snp.test.sqlite', tempdir())
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
x <- sqlite.build(filename = test.dat, list(dbname = test.sqlite, 
table.name = 'snp_test'))
sqlite.head(list(dbname = test.sqlite, table.name = 'snp_test'))
}
