% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{sqlite.index}
\alias{sqlite.index}
\title{Generate search index of annovarR database table in sqlite}
\usage{
sqlite.index(db, table.name, index, cols, verbose = FALSE)
}
\arguments{
\item{db}{Output path of sqlite database}

\item{table.name}{Table name in sqlite}

\item{index}{Index name in sqlite}

\item{cols}{Colnames needed to be index}

\item{verbose}{Ligical indicating wheather show the log message}

\item{...}{Other parameters be used in \code{\link[data.table]{fread}}}
}
\description{
Generate search index of annovarR database table in sqlite
}
\examples{
test.sqlite <- sprintf('\%s/snp.test.sqlite', tempdir())
test.dat <- system.file('extdata', 'demo/sqlite.dat.txt', package = 'annovarR')
x <- sqlite.build(filename = test.dat, db = test.sqlite, table.name = 'snp_test')
x <- sqlite.index(db = test.sqlite, table.name = 'snp_test', index = 'index4', cols = c('V1', 'V2'))
unlink(test.sqlite)
}
